requirements.txt
# Meta-Log Python Requirements
# Core dependencies for Python services and tests

# Testing and Coverage
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# FastAPI and web services
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Scientific computing (for services)
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0

# HTTP client
requests==2.31.0

